requirements/dev.txt
-r common.txt

GitPython
